Module Tezos_protocol_014_PtKathma.Protocol

val hash : Tezos_crypto.Hashed.Protocol_hash.t

val name : string

include module type of struct include Tezos_raw_protocol_014_PtKathma end

module Alpha_context = Tezos_raw_protocol_014_PtKathma.Alpha_context

module Alpha_services = Tezos_raw_protocol_014_PtKathma.Alpha_services

module Amendment = Tezos_raw_protocol_014_PtKathma.Amendment

module Apply = Tezos_raw_protocol_014_PtKathma.Apply

module Apply_internal_results = Tezos_raw_protocol_014_PtKathma.Apply_internal_results

module Apply_operation_result = Tezos_raw_protocol_014_PtKathma.Apply_operation_result

module Apply_results = Tezos_raw_protocol_014_PtKathma.Apply_results

module Baking = Tezos_raw_protocol_014_PtKathma.Baking

module Bitset = Tezos_raw_protocol_014_PtKathma.Bitset

module Blinded_public_key_hash = Tezos_raw_protocol_014_PtKathma.Blinded_public_key_hash

module Block_header_repr = Tezos_raw_protocol_014_PtKathma.Block_header_repr

module Block_payload_hash = Tezos_raw_protocol_014_PtKathma.Block_payload_hash

module Block_payload_repr = Tezos_raw_protocol_014_PtKathma.Block_payload_repr

module Bond_id_repr = Tezos_raw_protocol_014_PtKathma.Bond_id_repr

module Bootstrap_storage = Tezos_raw_protocol_014_PtKathma.Bootstrap_storage

module Cache_memory_helpers = Tezos_raw_protocol_014_PtKathma.Cache_memory_helpers

module Cache_repr = Tezos_raw_protocol_014_PtKathma.Cache_repr

module Carbonated_map = Tezos_raw_protocol_014_PtKathma.Carbonated_map

module Carbonated_map_costs = Tezos_raw_protocol_014_PtKathma.Carbonated_map_costs

module Commitment_repr = Tezos_raw_protocol_014_PtKathma.Commitment_repr

module Commitment_storage = Tezos_raw_protocol_014_PtKathma.Commitment_storage

module Constants_parametric_previous_repr = Tezos_raw_protocol_014_PtKathma.Constants_parametric_previous_repr

module Constants_parametric_repr = Tezos_raw_protocol_014_PtKathma.Constants_parametric_repr

module Constants_repr = Tezos_raw_protocol_014_PtKathma.Constants_repr

module Constants_services = Tezos_raw_protocol_014_PtKathma.Constants_services

module Constants_storage = Tezos_raw_protocol_014_PtKathma.Constants_storage

module Contract_delegate_storage = Tezos_raw_protocol_014_PtKathma.Contract_delegate_storage

module Contract_hash = Tezos_raw_protocol_014_PtKathma.Contract_hash

module Contract_manager_storage = Tezos_raw_protocol_014_PtKathma.Contract_manager_storage

module Contract_repr = Tezos_raw_protocol_014_PtKathma.Contract_repr

module Contract_services = Tezos_raw_protocol_014_PtKathma.Contract_services

module Contract_storage = Tezos_raw_protocol_014_PtKathma.Contract_storage

module Cycle_repr = Tezos_raw_protocol_014_PtKathma.Cycle_repr

module Dal_apply = Tezos_raw_protocol_014_PtKathma.Dal_apply

module Dal_endorsement_repr = Tezos_raw_protocol_014_PtKathma.Dal_endorsement_repr

module Dal_errors_repr = Tezos_raw_protocol_014_PtKathma.Dal_errors_repr

module Dal_slot_repr = Tezos_raw_protocol_014_PtKathma.Dal_slot_repr

module Dal_slot_storage = Tezos_raw_protocol_014_PtKathma.Dal_slot_storage

module Delegate_activation_storage = Tezos_raw_protocol_014_PtKathma.Delegate_activation_storage

module Delegate_services = Tezos_raw_protocol_014_PtKathma.Delegate_services

module Delegate_storage = Tezos_raw_protocol_014_PtKathma.Delegate_storage

module Dependent_bool = Tezos_raw_protocol_014_PtKathma.Dependent_bool

module Destination_repr = Tezos_raw_protocol_014_PtKathma.Destination_repr

module Entrypoint_repr = Tezos_raw_protocol_014_PtKathma.Entrypoint_repr

module Fees_storage = Tezos_raw_protocol_014_PtKathma.Fees_storage

module Fitness_repr = Tezos_raw_protocol_014_PtKathma.Fitness_repr

module Fixed_point_repr = Tezos_raw_protocol_014_PtKathma.Fixed_point_repr

module Frozen_deposits_storage = Tezos_raw_protocol_014_PtKathma.Frozen_deposits_storage

module Gas_comparable_input_size = Tezos_raw_protocol_014_PtKathma.Gas_comparable_input_size

module Gas_input_size = Tezos_raw_protocol_014_PtKathma.Gas_input_size

module Gas_limit_repr = Tezos_raw_protocol_014_PtKathma.Gas_limit_repr

module Gas_monad = Tezos_raw_protocol_014_PtKathma.Gas_monad

module Global_constants_costs = Tezos_raw_protocol_014_PtKathma.Global_constants_costs

module Global_constants_storage = Tezos_raw_protocol_014_PtKathma.Global_constants_storage

module Indexable = Tezos_raw_protocol_014_PtKathma.Indexable

module Init_storage = Tezos_raw_protocol_014_PtKathma.Init_storage

module Lazy_storage_diff = Tezos_raw_protocol_014_PtKathma.Lazy_storage_diff

module Lazy_storage_kind = Tezos_raw_protocol_014_PtKathma.Lazy_storage_kind

module Level_repr = Tezos_raw_protocol_014_PtKathma.Level_repr

module Level_storage = Tezos_raw_protocol_014_PtKathma.Level_storage

module Liquidity_baking_cpmm = Tezos_raw_protocol_014_PtKathma.Liquidity_baking_cpmm

module Liquidity_baking_lqt = Tezos_raw_protocol_014_PtKathma.Liquidity_baking_lqt

module Liquidity_baking_migration = Tezos_raw_protocol_014_PtKathma.Liquidity_baking_migration

module Liquidity_baking_repr = Tezos_raw_protocol_014_PtKathma.Liquidity_baking_repr

module Liquidity_baking_storage = Tezos_raw_protocol_014_PtKathma.Liquidity_baking_storage

module Local_gas_counter = Tezos_raw_protocol_014_PtKathma.Local_gas_counter

module Main = Tezos_raw_protocol_014_PtKathma.Main

module Manager_repr = Tezos_raw_protocol_014_PtKathma.Manager_repr

module Merkle_list = Tezos_raw_protocol_014_PtKathma.Merkle_list

module Michelson_v1_gas = Tezos_raw_protocol_014_PtKathma.Michelson_v1_gas

module Michelson_v1_primitives = Tezos_raw_protocol_014_PtKathma.Michelson_v1_primitives

module Migration_repr = Tezos_raw_protocol_014_PtKathma.Migration_repr

module Misc = Tezos_raw_protocol_014_PtKathma.Misc

module Non_empty_string = Tezos_raw_protocol_014_PtKathma.Non_empty_string

module Nonce_hash = Tezos_raw_protocol_014_PtKathma.Nonce_hash

module Nonce_storage = Tezos_raw_protocol_014_PtKathma.Nonce_storage

module Operation_repr = Tezos_raw_protocol_014_PtKathma.Operation_repr

module Origination_nonce = Tezos_raw_protocol_014_PtKathma.Origination_nonce

module Parameters_repr = Tezos_raw_protocol_014_PtKathma.Parameters_repr

module Path_encoding = Tezos_raw_protocol_014_PtKathma.Path_encoding

module Period_repr = Tezos_raw_protocol_014_PtKathma.Period_repr

module Ratio_repr = Tezos_raw_protocol_014_PtKathma.Ratio_repr

module Raw_context = Tezos_raw_protocol_014_PtKathma.Raw_context

module Raw_context_intf = Tezos_raw_protocol_014_PtKathma.Raw_context_intf

module Raw_level_repr = Tezos_raw_protocol_014_PtKathma.Raw_level_repr

module Receipt_repr = Tezos_raw_protocol_014_PtKathma.Receipt_repr

module Round_repr = Tezos_raw_protocol_014_PtKathma.Round_repr

module Sampler = Tezos_raw_protocol_014_PtKathma.Sampler

module Sapling_repr = Tezos_raw_protocol_014_PtKathma.Sapling_repr

module Sapling_services = Tezos_raw_protocol_014_PtKathma.Sapling_services

module Sapling_storage = Tezos_raw_protocol_014_PtKathma.Sapling_storage

module Sapling_validator = Tezos_raw_protocol_014_PtKathma.Sapling_validator

module Saturation_repr = Tezos_raw_protocol_014_PtKathma.Saturation_repr

module Sc_rollup_PVM_sem = Tezos_raw_protocol_014_PtKathma.Sc_rollup_PVM_sem

module Sc_rollup_arith = Tezos_raw_protocol_014_PtKathma.Sc_rollup_arith

module Sc_rollup_commitment_repr = Tezos_raw_protocol_014_PtKathma.Sc_rollup_commitment_repr

module Sc_rollup_commitment_storage = Tezos_raw_protocol_014_PtKathma.Sc_rollup_commitment_storage

module Sc_rollup_costs = Tezos_raw_protocol_014_PtKathma.Sc_rollup_costs

module Sc_rollup_data_version_sig = Tezos_raw_protocol_014_PtKathma.Sc_rollup_data_version_sig

module Sc_rollup_errors = Tezos_raw_protocol_014_PtKathma.Sc_rollup_errors

module Sc_rollup_game_repr = Tezos_raw_protocol_014_PtKathma.Sc_rollup_game_repr

module Sc_rollup_inbox_message_repr = Tezos_raw_protocol_014_PtKathma.Sc_rollup_inbox_message_repr

module Sc_rollup_inbox_repr = Tezos_raw_protocol_014_PtKathma.Sc_rollup_inbox_repr

module Sc_rollup_inbox_storage = Tezos_raw_protocol_014_PtKathma.Sc_rollup_inbox_storage

module Sc_rollup_management_protocol = Tezos_raw_protocol_014_PtKathma.Sc_rollup_management_protocol

module Sc_rollup_operations = Tezos_raw_protocol_014_PtKathma.Sc_rollup_operations

module Sc_rollup_outbox_message_repr = Tezos_raw_protocol_014_PtKathma.Sc_rollup_outbox_message_repr

module Sc_rollup_proof_repr = Tezos_raw_protocol_014_PtKathma.Sc_rollup_proof_repr

module Sc_rollup_refutation_storage = Tezos_raw_protocol_014_PtKathma.Sc_rollup_refutation_storage

module Sc_rollup_repr = Tezos_raw_protocol_014_PtKathma.Sc_rollup_repr

module Sc_rollup_stake_storage = Tezos_raw_protocol_014_PtKathma.Sc_rollup_stake_storage

module Sc_rollup_storage = Tezos_raw_protocol_014_PtKathma.Sc_rollup_storage

module Sc_rollup_tick_repr = Tezos_raw_protocol_014_PtKathma.Sc_rollup_tick_repr

module Sc_rollup_wasm = Tezos_raw_protocol_014_PtKathma.Sc_rollup_wasm

module Sc_rollups = Tezos_raw_protocol_014_PtKathma.Sc_rollups

module Script_big_map = Tezos_raw_protocol_014_PtKathma.Script_big_map

module Script_cache = Tezos_raw_protocol_014_PtKathma.Script_cache

module Script_comparable = Tezos_raw_protocol_014_PtKathma.Script_comparable

module Script_expr_hash = Tezos_raw_protocol_014_PtKathma.Script_expr_hash

module Script_int = Tezos_raw_protocol_014_PtKathma.Script_int

module Script_interpreter = Tezos_raw_protocol_014_PtKathma.Script_interpreter

module Script_interpreter_defs = Tezos_raw_protocol_014_PtKathma.Script_interpreter_defs

module Script_interpreter_logging = Tezos_raw_protocol_014_PtKathma.Script_interpreter_logging

module Script_ir_annot = Tezos_raw_protocol_014_PtKathma.Script_ir_annot

module Script_ir_translator = Tezos_raw_protocol_014_PtKathma.Script_ir_translator

module Script_list = Tezos_raw_protocol_014_PtKathma.Script_list

module Script_map = Tezos_raw_protocol_014_PtKathma.Script_map

module Script_repr = Tezos_raw_protocol_014_PtKathma.Script_repr

module Script_set = Tezos_raw_protocol_014_PtKathma.Script_set

module Script_string = Tezos_raw_protocol_014_PtKathma.Script_string

module Script_tc_context = Tezos_raw_protocol_014_PtKathma.Script_tc_context

module Script_tc_errors = Tezos_raw_protocol_014_PtKathma.Script_tc_errors

module Script_tc_errors_registration = Tezos_raw_protocol_014_PtKathma.Script_tc_errors_registration

module Script_timestamp = Tezos_raw_protocol_014_PtKathma.Script_timestamp

module Script_typed_ir = Tezos_raw_protocol_014_PtKathma.Script_typed_ir

module Script_typed_ir_size = Tezos_raw_protocol_014_PtKathma.Script_typed_ir_size

module Script_typed_ir_size_costs = Tezos_raw_protocol_014_PtKathma.Script_typed_ir_size_costs

module Seed_repr = Tezos_raw_protocol_014_PtKathma.Seed_repr

module Seed_storage = Tezos_raw_protocol_014_PtKathma.Seed_storage

module Services_registration = Tezos_raw_protocol_014_PtKathma.Services_registration

module Skip_list_repr = Tezos_raw_protocol_014_PtKathma.Skip_list_repr

module Slot_repr = Tezos_raw_protocol_014_PtKathma.Slot_repr

module Stake_storage = Tezos_raw_protocol_014_PtKathma.Stake_storage

module State_hash = Tezos_raw_protocol_014_PtKathma.State_hash

module Storage = Tezos_raw_protocol_014_PtKathma.Storage

module Storage_costs = Tezos_raw_protocol_014_PtKathma.Storage_costs

module Storage_description = Tezos_raw_protocol_014_PtKathma.Storage_description

module Storage_functors = Tezos_raw_protocol_014_PtKathma.Storage_functors

module Storage_sigs = Tezos_raw_protocol_014_PtKathma.Storage_sigs

module Tez_repr = Tezos_raw_protocol_014_PtKathma.Tez_repr

module Ticket_accounting = Tezos_raw_protocol_014_PtKathma.Ticket_accounting

module Ticket_balance_key = Tezos_raw_protocol_014_PtKathma.Ticket_balance_key

module Ticket_costs = Tezos_raw_protocol_014_PtKathma.Ticket_costs

module Ticket_hash_builder = Tezos_raw_protocol_014_PtKathma.Ticket_hash_builder

module Ticket_hash_repr = Tezos_raw_protocol_014_PtKathma.Ticket_hash_repr

module Ticket_lazy_storage_diff = Tezos_raw_protocol_014_PtKathma.Ticket_lazy_storage_diff

module Ticket_operations_diff = Tezos_raw_protocol_014_PtKathma.Ticket_operations_diff

module Ticket_scanner = Tezos_raw_protocol_014_PtKathma.Ticket_scanner

module Ticket_storage = Tezos_raw_protocol_014_PtKathma.Ticket_storage

module Ticket_token = Tezos_raw_protocol_014_PtKathma.Ticket_token

module Ticket_token_map = Tezos_raw_protocol_014_PtKathma.Ticket_token_map

module Time_repr = Tezos_raw_protocol_014_PtKathma.Time_repr

module Token = Tezos_raw_protocol_014_PtKathma.Token

module Tx_rollup_commitment_repr = Tezos_raw_protocol_014_PtKathma.Tx_rollup_commitment_repr

module Tx_rollup_commitment_storage = Tezos_raw_protocol_014_PtKathma.Tx_rollup_commitment_storage

module Tx_rollup_errors_repr = Tezos_raw_protocol_014_PtKathma.Tx_rollup_errors_repr

module Tx_rollup_gas = Tezos_raw_protocol_014_PtKathma.Tx_rollup_gas

module Tx_rollup_hash_builder = Tezos_raw_protocol_014_PtKathma.Tx_rollup_hash_builder

module Tx_rollup_inbox_repr = Tezos_raw_protocol_014_PtKathma.Tx_rollup_inbox_repr

module Tx_rollup_inbox_storage = Tezos_raw_protocol_014_PtKathma.Tx_rollup_inbox_storage

module Tx_rollup_l2_address = Tezos_raw_protocol_014_PtKathma.Tx_rollup_l2_address

module Tx_rollup_l2_apply = Tezos_raw_protocol_014_PtKathma.Tx_rollup_l2_apply

module Tx_rollup_l2_batch = Tezos_raw_protocol_014_PtKathma.Tx_rollup_l2_batch

module Tx_rollup_l2_context = Tezos_raw_protocol_014_PtKathma.Tx_rollup_l2_context

module Tx_rollup_l2_context_hash = Tezos_raw_protocol_014_PtKathma.Tx_rollup_l2_context_hash

module Tx_rollup_l2_context_sig = Tezos_raw_protocol_014_PtKathma.Tx_rollup_l2_context_sig

module Tx_rollup_l2_proof = Tezos_raw_protocol_014_PtKathma.Tx_rollup_l2_proof

module Tx_rollup_l2_qty = Tezos_raw_protocol_014_PtKathma.Tx_rollup_l2_qty

module Tx_rollup_l2_storage_sig = Tezos_raw_protocol_014_PtKathma.Tx_rollup_l2_storage_sig

module Tx_rollup_l2_verifier = Tezos_raw_protocol_014_PtKathma.Tx_rollup_l2_verifier

module Tx_rollup_level_repr = Tezos_raw_protocol_014_PtKathma.Tx_rollup_level_repr

module Tx_rollup_message_hash_repr = Tezos_raw_protocol_014_PtKathma.Tx_rollup_message_hash_repr

module Tx_rollup_message_repr = Tezos_raw_protocol_014_PtKathma.Tx_rollup_message_repr

module Tx_rollup_message_result_hash_repr = Tezos_raw_protocol_014_PtKathma.Tx_rollup_message_result_hash_repr

module Tx_rollup_message_result_repr = Tezos_raw_protocol_014_PtKathma.Tx_rollup_message_result_repr

module Tx_rollup_parameters = Tezos_raw_protocol_014_PtKathma.Tx_rollup_parameters

module Tx_rollup_prefixes = Tezos_raw_protocol_014_PtKathma.Tx_rollup_prefixes

module Tx_rollup_repr = Tezos_raw_protocol_014_PtKathma.Tx_rollup_repr

module Tx_rollup_reveal_repr = Tezos_raw_protocol_014_PtKathma.Tx_rollup_reveal_repr

module Tx_rollup_reveal_storage = Tezos_raw_protocol_014_PtKathma.Tx_rollup_reveal_storage

module Tx_rollup_services = Tezos_raw_protocol_014_PtKathma.Tx_rollup_services

module Tx_rollup_state_repr = Tezos_raw_protocol_014_PtKathma.Tx_rollup_state_repr

module Tx_rollup_state_storage = Tezos_raw_protocol_014_PtKathma.Tx_rollup_state_storage

module Tx_rollup_storage = Tezos_raw_protocol_014_PtKathma.Tx_rollup_storage

module Tx_rollup_ticket = Tezos_raw_protocol_014_PtKathma.Tx_rollup_ticket

module Tx_rollup_withdraw_list_hash_repr = Tezos_raw_protocol_014_PtKathma.Tx_rollup_withdraw_list_hash_repr

module Tx_rollup_withdraw_repr = Tezos_raw_protocol_014_PtKathma.Tx_rollup_withdraw_repr

module Validate_operation = Tezos_raw_protocol_014_PtKathma.Validate_operation

module Vote_repr = Tezos_raw_protocol_014_PtKathma.Vote_repr

module Vote_storage = Tezos_raw_protocol_014_PtKathma.Vote_storage

module Voting_period_repr = Tezos_raw_protocol_014_PtKathma.Voting_period_repr

module Voting_period_storage = Tezos_raw_protocol_014_PtKathma.Voting_period_storage

module Voting_services = Tezos_raw_protocol_014_PtKathma.Voting_services

include module type of struct include Tezos_raw_protocol_014_PtKathma.Main end

type validation_mode = Tezos_raw_protocol_014_PtKathma.Main.validation_mode =

| Application of {

block_header : Tezos_raw_protocol_014_PtKathma.Alpha_context.Block_header.t;

fitness : Tezos_raw_protocol_014_PtKathma.Alpha_context.Fitness.t;

payload_producer : Tezos_raw_protocol_014_PtKathma.Alpha_context.public_key_hash;

block_producer : Tezos_raw_protocol_014_PtKathma.Alpha_context.public_key_hash;

predecessor_round : Tezos_raw_protocol_014_PtKathma.Alpha_context.Round.t;

predecessor_level : Tezos_raw_protocol_014_PtKathma.Alpha_context.Level.t;

}
(*
Full Validation of a block. See Tezos_protocol_environment_sigs.V5.T.Updater.PROTOCOL.begin_application*
*)

| Partial_application of {

block_header : Tezos_raw_protocol_014_PtKathma.Alpha_context.Block_header.t;

fitness : Tezos_raw_protocol_014_PtKathma.Alpha_context.Fitness.t;

payload_producer : Tezos_raw_protocol_014_PtKathma.Alpha_context.public_key_hash;

block_producer : Tezos_raw_protocol_014_PtKathma.Alpha_context.public_key_hash;

predecessor_level : Tezos_raw_protocol_014_PtKathma.Alpha_context.Level.t;

predecessor_round : Tezos_raw_protocol_014_PtKathma.Alpha_context.Round.t;

}
(*
Partial_application is used in pre-checking of blocks - not all checks are done. Special case of Application to allow quick rejection of bad blocks. See Tezos_protocol_environment_sigs.V5.T.Updater.PROTOCOL.begin_partial_application
*)

| Partial_construction of {

predecessor : Tezos_protocol_environment_014_PtKathma.Block_hash.t;

predecessor_fitness : Tezos_protocol_environment_014_PtKathma.Fitness.t;

predecessor_level : Tezos_raw_protocol_014_PtKathma.Alpha_context.Level.t;

predecessor_round : Tezos_raw_protocol_014_PtKathma.Alpha_context.Round.t;

}
(*
Shell/mempool-only construction of a virtual block. See Tezos_protocol_environment_sigs.V5.T.Updater.PROTOCOL.begin_construction
*)

| Full_construction of {

predecessor : Tezos_protocol_environment_014_PtKathma.Block_hash.t;

payload_producer : Tezos_raw_protocol_014_PtKathma.Alpha_context.public_key_hash;

block_producer : Tezos_raw_protocol_014_PtKathma.Alpha_context.public_key_hash;

protocol_data_contents : Tezos_raw_protocol_014_PtKathma.Alpha_context.Block_header.contents;

level : Tezos_protocol_environment_014_PtKathma.Int32.t;

round : Tezos_raw_protocol_014_PtKathma.Alpha_context.Round.t;

predecessor_level : Tezos_raw_protocol_014_PtKathma.Alpha_context.Level.t;

predecessor_round : Tezos_raw_protocol_014_PtKathma.Alpha_context.Round.t;

}
(*
Baker-only block construction for baking in. See Tezos_protocol_environment_sigs.V5.T.Updater.PROTOCOL.begin_construction
*)

validation_mode permits to differenciate !type:validation_state values.

There are four validation modes:

Application

Partial_application

Partial_construction

Full_construction

For the meaning and typical uses of each mode, refer to the comments attached to the corresponding type constructors below.

type validation_state = Tezos_raw_protocol_014_PtKathma.Main.validation_state = {

mode : validation_mode;

chain_id : Tezos_protocol_environment_014_PtKathma.Chain_id.t;

ctxt : Tezos_raw_protocol_014_PtKathma.Alpha_context.t;

op_count : int;

migration_balance_updates : Tezos_raw_protocol_014_PtKathma.Alpha_context.Receipt.balance_updates;

liquidity_baking_toggle_ema : Tezos_raw_protocol_014_PtKathma.Alpha_context.Liquidity_baking.Toggle_EMA.t;

implicit_operations_results : Tezos_raw_protocol_014_PtKathma.Apply_results.packed_successful_manager_operation_result list;

validate_operation_info : Tezos_raw_protocol_014_PtKathma.Validate_operation.validate_operation_info;

validate_operation_state : Tezos_raw_protocol_014_PtKathma.Validate_operation.validate_operation_state;

}

type operation_data = Tezos_raw_protocol_014_PtKathma.Alpha_context.packed_protocol_data

type operation = Tezos_raw_protocol_014_PtKathma.Alpha_context.packed_operation = {

shell : Tezos_protocol_environment_014_PtKathma.Operation.shell_header;

protocol_data : operation_data;

}

val precheck_manager : validation_state -> 'a Tezos_raw_protocol_014_PtKathma.Alpha_context.Kind.manager Tezos_raw_protocol_014_PtKathma.Alpha_context.contents_list -> 'a Tezos_raw_protocol_014_PtKathma.Alpha_context.Kind.manager Tezos_raw_protocol_014_PtKathma.Validate_operation.TMP_for_plugin.should_check_signature -> Tezos_raw_protocol_014_PtKathma.Validate_operation.stamp Tezos_protocol_environment_014_PtKathma.Error_monad.tzresult Tezos_protocol_environment_014_PtKathma.Lwt.t

See Validate_operation.precheck_manager.

include Tezos_protocol_environment_014_PtKathma.Updater.PROTOCOL with type block_header_data = Tezos_raw_protocol_014_PtKathma.Alpha_context.Block_header.protocol_data and type block_header_metadata = Tezos_raw_protocol_014_PtKathma.Apply_results.block_metadata and type block_header = Tezos_raw_protocol_014_PtKathma.Alpha_context.Block_header.t and type operation_data := operation_data and type operation_receipt = Tezos_raw_protocol_014_PtKathma.Apply_results.packed_operation_metadata and type operation := operation and type validation_state := validation_state

val max_block_length : int

val max_operation_data_length : int

val validation_passes : Tezos_protocol_environment_014_PtKathma.Updater.quota list

type block_header_data = Tezos_raw_protocol_014_PtKathma.Alpha_context.Block_header.protocol_data

val block_header_data_encoding : block_header_data Tezos_protocol_environment_014_PtKathma.Data_encoding.t

type block_header = Tezos_raw_protocol_014_PtKathma.Alpha_context.Block_header.t = {

shell : Tezos_protocol_environment_014_PtKathma.Block_header.shell_header;

protocol_data : block_header_data;

}

type block_header_metadata = Tezos_raw_protocol_014_PtKathma.Apply_results.block_metadata

val block_header_metadata_encoding : block_header_metadata Tezos_protocol_environment_014_PtKathma.Data_encoding.t

type operation_receipt = Tezos_raw_protocol_014_PtKathma.Apply_results.packed_operation_metadata

val operation_data_encoding : operation_data Tezos_protocol_environment_014_PtKathma.Data_encoding.t

val operation_receipt_encoding : operation_receipt Tezos_protocol_environment_014_PtKathma.Data_encoding.t

val operation_data_and_receipt_encoding : (operation_data * operation_receipt) Tezos_protocol_environment_014_PtKathma.Data_encoding.t

val acceptable_passes : operation -> int list

val relative_position_within_block : operation -> operation -> int

val begin_partial_application : chain_id:Tezos_protocol_environment_014_PtKathma.Chain_id.t -> ancestor_context:Tezos_protocol_environment_014_PtKathma.Context.t -> predecessor_timestamp:Tezos_protocol_environment_014_PtKathma.Time.t -> predecessor_fitness:Tezos_protocol_environment_014_PtKathma.Fitness.t -> block_header -> validation_state Tezos_protocol_environment_014_PtKathma.Error_monad.tzresult Tezos_protocol_environment_014_PtKathma.Lwt.t

val begin_application : chain_id:Tezos_protocol_environment_014_PtKathma.Chain_id.t -> predecessor_context:Tezos_protocol_environment_014_PtKathma.Context.t -> predecessor_timestamp:Tezos_protocol_environment_014_PtKathma.Time.t -> predecessor_fitness:Tezos_protocol_environment_014_PtKathma.Fitness.t -> block_header -> validation_state Tezos_protocol_environment_014_PtKathma.Error_monad.tzresult Tezos_protocol_environment_014_PtKathma.Lwt.t

val begin_construction : chain_id:Tezos_protocol_environment_014_PtKathma.Chain_id.t -> predecessor_context:Tezos_protocol_environment_014_PtKathma.Context.t -> predecessor_timestamp:Tezos_protocol_environment_014_PtKathma.Time.t -> predecessor_level:Tezos_protocol_environment_014_PtKathma.Int32.t -> predecessor_fitness:Tezos_protocol_environment_014_PtKathma.Fitness.t -> predecessor:Tezos_protocol_environment_014_PtKathma.Block_hash.t -> timestamp:Tezos_protocol_environment_014_PtKathma.Time.t -> ?protocol_data:block_header_data -> unit -> validation_state Tezos_protocol_environment_014_PtKathma.Error_monad.tzresult Tezos_protocol_environment_014_PtKathma.Lwt.t

val apply_operation : validation_state -> operation -> (validation_state * operation_receipt) Tezos_protocol_environment_014_PtKathma.Error_monad.tzresult Tezos_protocol_environment_014_PtKathma.Lwt.t

val finalize_block : validation_state -> Tezos_protocol_environment_014_PtKathma.Block_header.shell_header option -> (Tezos_protocol_environment_014_PtKathma.Updater.validation_result * block_header_metadata) Tezos_protocol_environment_014_PtKathma.Error_monad.tzresult Tezos_protocol_environment_014_PtKathma.Lwt.t

val rpc_services : Tezos_protocol_environment_014_PtKathma.Updater.rpc_context Tezos_protocol_environment_014_PtKathma.RPC_directory.t

val init : Tezos_protocol_environment_014_PtKathma.Chain_id.t -> Tezos_protocol_environment_014_PtKathma.Context.t -> Tezos_protocol_environment_014_PtKathma.Block_header.shell_header -> Tezos_protocol_environment_014_PtKathma.Updater.validation_result Tezos_protocol_environment_014_PtKathma.Error_monad.tzresult Tezos_protocol_environment_014_PtKathma.Lwt.t

val value_of_key : chain_id:Tezos_protocol_environment_014_PtKathma.Chain_id.t -> predecessor_context:Tezos_protocol_environment_014_PtKathma.Context.t -> predecessor_timestamp:Tezos_protocol_environment_014_PtKathma.Time.t -> predecessor_level:Tezos_protocol_environment_014_PtKathma.Int32.t -> predecessor_fitness:Tezos_protocol_environment_014_PtKathma.Fitness.t -> predecessor:Tezos_protocol_environment_014_PtKathma.Block_hash.t -> timestamp:Tezos_protocol_environment_014_PtKathma.Time.t -> (Tezos_protocol_environment_014_PtKathma.Context.Cache.key -> Tezos_protocol_environment_014_PtKathma.Context.Cache.value Tezos_protocol_environment_014_PtKathma.Error_monad.tzresult Tezos_protocol_environment_014_PtKathma.Lwt.t) Tezos_protocol_environment_014_PtKathma.Error_monad.tzresult Tezos_protocol_environment_014_PtKathma.Lwt.t