Module Tezos_raw_protocol_014_PtKathma.Alpha_context

An Alpha_context.t is an immutable snapshot of the ledger state at some block height, preserving type-safety and invariants of the ledger state.

Implementation

Alpha_context.t is a wrapper over Raw_context.t, which in turn is a wrapper around Context.t from the Protocol Environment.

Lifetime of an Alpha_context

Creation, using prepare or prepare_first_block

Modification, using the operations defined in this signature

Finalization, using finalize

module type BASIC_DATA = sig ... end

type t

type context = t

type public_key = Tezos_protocol_environment_014_PtKathma.Signature.Public_key.t

type public_key_hash = Tezos_protocol_environment_014_PtKathma.Signature.Public_key_hash.t

type signature = Tezos_protocol_environment_014_PtKathma.Signature.t

module Slot : sig ... end

This module re-exports definitions from Slot_repr.

module Tez : sig ... end

This module re-exports definitions from Tez_repr.

module Period : sig ... end

This module re-exports definitions from Period_repr.

module Timestamp : sig ... end

This module re-exports definitions from Time_repr.

module Ratio : sig ... end

This module re-exports definitions from Ratio_repr.

module Raw_level : sig ... end

This module re-exports definitions from Raw_level_repr.

module Cycle : sig ... end

This module re-exports definitions from Cycle_repr.

module Round : sig ... end

This module re-exports definitions from Round_repr.

module Gas : sig ... end

This module implements the gas subsystem of the context.

module Entrypoint : module type of Entrypoint_repr

module Script : sig ... end

This module re-exports definitions from Script_repr and Michelson_v1_primitives.

module Constants : sig ... end

This module re-exports definitions from Constants_repr and Constants_storage.

module Global_constants_storage : sig ... end

See the definitions inside the module.

module Internal_for_tests : sig ... end

This module discloses definitions that are only useful for tests and must not be used otherwise.

module Cache : sig ... end

This module re-exports definitions from Cache_repr.

module Level : sig ... end

This module re-exports definitions from Level_repr and Level_storage.

module Fitness : sig ... end

This module re-exports definitions from Fitness_repr.

module Nonce : sig ... end

This module re-exports definitions from Nonce_storage.

module Seed : sig ... end

This module re-exports definitions from Seed_repr and Seed_storage.

module Big_map : sig ... end

module Sapling : sig ... end

This module re-exports definitions from Sapling_repr, Sapling_storage and Sapling_validator.

module Lazy_storage : sig ... end

This module re-exports definitions from Lazy_storage_diff.

module Origination_nonce : sig ... end

See the definitions inside the module.

module Ticket_hash : sig ... end

This module re-exports definitions from Ticket_hash_repr.

module Contract : sig ... end

This module re-exports definitions from Contract_repr and Contract_storage.

module Tx_rollup_level : sig ... end

This module re-exports definitions from Tx_rollup_level_repr.

module Tx_rollup : sig ... end

This module re-exports definitions from Tx_rollup_repr and Tx_rollup_storage.

module Tx_rollup_withdraw : sig ... end

This module re-exports definitions from Tx_rollup_withdraw_repr.

module Tx_rollup_withdraw_list_hash : sig ... end

This module re-exports definitions from Tx_rollup_withdraw_list_hash_repr.

module Tx_rollup_message_result : sig ... end

This module re-exports definitions from Tx_rollup_message_result_repr.

module Tx_rollup_message_result_hash : sig ... end

This module re-exports definitions from Tx_rollup_message_result_hash_repr.

module Tx_rollup_commitment_hash : sig ... end

This module re-exports definitions from Tx_rollup_commitment_repr.Hash.

module Tx_rollup_state : sig ... end

This module re-exports definitions from Tx_rollup_state_repr and Tx_rollup_state_storage.

module Tx_rollup_reveal : sig ... end

This module re-exports definitions from Tx_rollup_reveal_repr and Tx_rollup_reveal_storage.

module Tx_rollup_message : sig ... end

This module re-exports definitions from Tx_rollup_message_repr.

module Tx_rollup_message_hash : sig ... end

This module re-exports definitions from Tx_rollup_message_hash_repr.

module Tx_rollup_inbox : sig ... end

This module re-exports definitions from Tx_rollup_inbox_repr and Tx_rollup_inbox_storage.

module Tx_rollup_commitment : sig ... end

This module re-exports definitions from Tx_rollup_commitment_repr.

module Tx_rollup_hash : sig ... end

This module re-exports definitions from Tx_rollup_hash_builder.

module Tx_rollup_errors : sig ... end

This module re-exports definitions from Tx_rollup_errors_repr.

module Sc_rollup_repr : sig ... end

This is a forward declaration to avoid circular dependencies. Use module Sc_rollup instead whenever possible. TODO : find a better way to resolve the circular dependency https://gitlab.com/tezos/tezos/-/issues/3147

module Bond_id : sig ... end

This module re-exports definitions from Bond_id_repr.

module Receipt : sig ... end

This module re-exports definitions from Receipt_repr.

module Delegate : sig ... end

This module re-exports definitions from Delegate_storage.

module Voting_period : sig ... end

This module re-exports definitions from Voting_period_repr and Voting_period_storage.

module Vote : sig ... end

This module re-exports definitions from Vote_repr and Vote_storage.

module Dal : sig ... end

module Dal_errors : sig ... end

module Sc_rollup : sig ... end

This module re-exports definitions from Sc_rollup_storage and Sc_rollup_repr.

module Destination : sig ... end

This module re-exports definitions from Destination_repr.

module Block_payload : sig ... end

module Block_header : sig ... end

This module re-exports definitions from Block_header_repr.

module Kind : sig ... end

This module re-exports definitions from Lazy_storage_kind.

All the definitions below are re-exported from Operation_repr.

type 'a consensus_operation_type =

| Endorsement : Kind.endorsement consensus_operation_type

| Preendorsement : Kind.preendorsement consensus_operation_type

val pp_operation_kind : Tezos_protocol_environment_014_PtKathma.Format.formatter -> 'kind consensus_operation_type -> unit

type consensus_content = {

slot : Slot.t;

level : Raw_level.t;

round : Round.t;

block_payload_hash : Block_payload_hash.t;

}

val consensus_content_encoding : consensus_content Tezos_protocol_environment_014_PtKathma.Data_encoding.t

val pp_consensus_content : Tezos_protocol_environment_014_PtKathma.Format.formatter -> consensus_content -> unit

type 'kind operation = {

shell : Tezos_protocol_environment_014_PtKathma.Operation.shell_header;

protocol_data : 'kind protocol_data;

}

and 'kind protocol_data = {

contents : 'kind contents_list;

signature : Tezos_protocol_environment_014_PtKathma.Signature.t option;

}

and _ contents_list =

| Single : 'kind contents -> 'kind contents_list

| Cons : 'kind Kind.manager contents * 'rest Kind.manager contents_list -> ('kind * 'rest) Kind.manager contents_list

and _ contents =

| Preendorsement : consensus_content -> Kind.preendorsement contents

| Endorsement : consensus_content -> Kind.endorsement contents

| Dal_slot_availability : Tezos_protocol_environment_014_PtKathma.Signature.Public_key_hash.t * Dal.Endorsement.t -> Kind.dal_slot_availability contents

| Seed_nonce_revelation : {

level : Raw_level.t;

nonce : Nonce.t;

} -> Kind.seed_nonce_revelation contents

| Vdf_revelation : {

solution : Seed.vdf_solution;

} -> Kind.vdf_revelation contents

| Double_preendorsement_evidence : {

op1 : Kind.preendorsement operation;

op2 : Kind.preendorsement operation;

} -> Kind.double_preendorsement_evidence contents

| Double_endorsement_evidence : {

op1 : Kind.endorsement operation;

op2 : Kind.endorsement operation;

} -> Kind.double_endorsement_evidence contents

| Double_baking_evidence : {

bh1 : Block_header.t;

bh2 : Block_header.t;

} -> Kind.double_baking_evidence contents

| Activate_account : {

id : Tezos_protocol_environment_014_PtKathma.Ed25519.Public_key_hash.t;

activation_code : Blinded_public_key_hash.activation_code;

} -> Kind.activate_account contents

| Proposals : {

source : Tezos_protocol_environment_014_PtKathma.Signature.Public_key_hash.t;

period : int32;

proposals : Tezos_protocol_environment_014_PtKathma.Protocol_hash.t list;

} -> Kind.proposals contents

| Ballot : {

source : Tezos_protocol_environment_014_PtKathma.Signature.Public_key_hash.t;

period : int32;

proposal : Tezos_protocol_environment_014_PtKathma.Protocol_hash.t;

ballot : Vote.ballot;

} -> Kind.ballot contents

| Failing_noop : string -> Kind.failing_noop contents

| Manager_operation : {

source : Tezos_protocol_environment_014_PtKathma.Signature.Public_key_hash.t;

fee : Tez.tez;

counter : counter;

operation : 'kind manager_operation;

gas_limit : Gas.Arith.integral;

storage_limit : Tezos_protocol_environment_014_PtKathma.Z.t;

} -> 'kind Kind.manager contents

and _ manager_operation =

| Reveal : Tezos_protocol_environment_014_PtKathma.Signature.Public_key.t -> Kind.reveal manager_operation

| Transaction : {

amount : Tez.tez;

parameters : Script.lazy_expr;

entrypoint : Entrypoint.t;

destination : Contract.t;

} -> Kind.transaction manager_operation

| Origination : {

delegate : Tezos_protocol_environment_014_PtKathma.Signature.Public_key_hash.t option;

script : Script.t;

credit : Tez.tez;

} -> Kind.origination manager_operation

| Delegation : Tezos_protocol_environment_014_PtKathma.Signature.Public_key_hash.t option -> Kind.delegation manager_operation

| Register_global_constant : {

value : Script.lazy_expr;

} -> Kind.register_global_constant manager_operation

| Set_deposits_limit : Tez.t option -> Kind.set_deposits_limit manager_operation

| Increase_paid_storage : {

amount_in_bytes : Tezos_protocol_environment_014_PtKathma.Z.t;

destination : Contract_hash.t;

} -> Kind.increase_paid_storage manager_operation

| Tx_rollup_origination : Kind.tx_rollup_origination manager_operation

| Tx_rollup_submit_batch : {

tx_rollup : Tx_rollup.t;

content : string;

burn_limit : Tez.tez option;

} -> Kind.tx_rollup_submit_batch manager_operation

| Tx_rollup_commit : {

tx_rollup : Tx_rollup.t;

commitment : Tx_rollup_commitment.Full.t;

} -> Kind.tx_rollup_commit manager_operation

| Tx_rollup_return_bond : {

tx_rollup : Tx_rollup.t;

} -> Kind.tx_rollup_return_bond manager_operation

| Tx_rollup_finalize_commitment : {

tx_rollup : Tx_rollup.t;

} -> Kind.tx_rollup_finalize_commitment manager_operation

| Tx_rollup_remove_commitment : {

tx_rollup : Tx_rollup.t;

} -> Kind.tx_rollup_remove_commitment manager_operation

| Tx_rollup_rejection : {

tx_rollup : Tx_rollup.t;

level : Tx_rollup_level.t;

message : Tx_rollup_message.t;

message_position : int;

message_path : Tx_rollup_inbox.Merkle.path;

message_result_hash : Tx_rollup_message_result_hash.t;

message_result_path : Tx_rollup_commitment.Merkle.path;

previous_message_result : Tx_rollup_message_result.t;

previous_message_result_path : Tx_rollup_commitment.Merkle.path;

proof : Tx_rollup_l2_proof.t;

} -> Kind.tx_rollup_rejection manager_operation

| Tx_rollup_dispatch_tickets : {

tx_rollup : Tx_rollup.t;

level : Tx_rollup_level.t;

context_hash : Tezos_protocol_environment_014_PtKathma.Context_hash.t;

message_index : int;

message_result_path : Tx_rollup_commitment.Merkle.path;

tickets_info : Tx_rollup_reveal.t list;

} -> Kind.tx_rollup_dispatch_tickets manager_operation

| Transfer_ticket : {

contents : Script.lazy_expr;

ty : Script.lazy_expr;

ticketer : Contract.t;

amount : Tezos_protocol_environment_014_PtKathma.Z.t;

destination : Contract.t;

entrypoint : Entrypoint.t;

} -> Kind.transfer_ticket manager_operation

| Dal_publish_slot_header : {

slot : Dal.Slot.t;

} -> Kind.dal_publish_slot_header manager_operation

| Sc_rollup_originate : {

kind : Sc_rollup.Kind.t;

boot_sector : string;

parameters_ty : Script.lazy_expr;

} -> Kind.sc_rollup_originate manager_operation

| Sc_rollup_add_messages : {

rollup : Sc_rollup.t;

messages : string list;

} -> Kind.sc_rollup_add_messages manager_operation

| Sc_rollup_cement : {

rollup : Sc_rollup.t;

commitment : Sc_rollup.Commitment.Hash.t;

} -> Kind.sc_rollup_cement manager_operation

| Sc_rollup_publish : {

rollup : Sc_rollup.t;

commitment : Sc_rollup.Commitment.t;

} -> Kind.sc_rollup_publish manager_operation

| Sc_rollup_refute : {

rollup : Sc_rollup.t;

opponent : Sc_rollup.Staker.t;

refutation : Sc_rollup.Game.refutation;

is_opening_move : bool;

} -> Kind.sc_rollup_refute manager_operation

| Sc_rollup_timeout : {

rollup : Sc_rollup.t;

stakers : Sc_rollup.Game.Index.t;

} -> Kind.sc_rollup_timeout manager_operation

| Sc_rollup_execute_outbox_message : {

rollup : Sc_rollup.t;

cemented_commitment : Sc_rollup.Commitment.Hash.t;

outbox_level : Raw_level.t;

message_index : int;

inclusion_proof : string;

message : string;

} -> Kind.sc_rollup_execute_outbox_message manager_operation

| Sc_rollup_recover_bond : {

sc_rollup : Sc_rollup.t;

} -> Kind.sc_rollup_recover_bond manager_operation

| Sc_rollup_dal_slot_subscribe : {

rollup : Sc_rollup.t;

slot_index : Dal.Slot_index.t;

} -> Kind.sc_rollup_dal_slot_subscribe manager_operation

and counter = Tezos_protocol_environment_014_PtKathma.Z.t

type packed_manager_operation =

| Manager : 'kind manager_operation -> packed_manager_operation

type packed_contents =

| Contents : 'kind contents -> packed_contents

type packed_contents_list =

| Contents_list : 'kind contents_list -> packed_contents_list

type packed_protocol_data =

| Operation_data : 'kind protocol_data -> packed_protocol_data

type packed_operation = {

shell : Tezos_protocol_environment_014_PtKathma.Operation.shell_header;

protocol_data : packed_protocol_data;

}

val manager_kind : 'kind manager_operation -> 'kind Kind.manager

module Operation : sig ... end

This module re-exports definitions from Operation_repr.

module Stake_distribution : sig ... end

This module re-exports definitions from Stake_storage, Delegate_storage and Delegate.

module Commitment : sig ... end

This module re-exports definitions from Commitment_repr and, Commitment_storage.

module Bootstrap : sig ... end

This module re-exports definitions from Bootstrap_storage.

module Migration : sig ... end

This module re-exports definitions from Migration_repr.

val prepare_first_block : Tezos_protocol_environment_014_PtKathma.Chain_id.t -> Tezos_protocol_environment_014_PtKathma.Context.t -> typecheck: (context -> Script.t -> ((Script.t * Lazy_storage.diffs option) * context) Tezos_protocol_environment_014_PtKathma.Error_monad.tzresult Tezos_protocol_environment_014_PtKathma.Lwt.t) -> level:Tezos_protocol_environment_014_PtKathma.Int32.t -> timestamp:Tezos_protocol_environment_014_PtKathma.Time.t -> context Tezos_protocol_environment_014_PtKathma.Error_monad.tzresult Tezos_protocol_environment_014_PtKathma.Lwt.t

Create an Alpha_context.t from an untyped context (first block in the chain only).

val prepare : Tezos_protocol_environment_014_PtKathma.Context.t -> level:Tezos_protocol_environment_014_PtKathma.Int32.t -> predecessor_timestamp:Tezos_protocol_environment_014_PtKathma.Time.t -> timestamp:Tezos_protocol_environment_014_PtKathma.Time.t -> (context * Receipt.balance_updates * Migration.origination_result list) Tezos_protocol_environment_014_PtKathma.Error_monad.tzresult Tezos_protocol_environment_014_PtKathma.Lwt.t

Create an Alpha_context.t from an untyped context.

All the definitions below are re-exported from Raw_context.

val activate : context -> Tezos_protocol_environment_014_PtKathma.Protocol_hash.t -> context Tezos_protocol_environment_014_PtKathma.Lwt.t

val reset_internal_nonce : context -> context

val fresh_internal_nonce : context -> (context * int) Tezos_protocol_environment_014_PtKathma.Error_monad.tzresult

val record_internal_nonce : context -> int -> context

val internal_nonce_already_recorded : context -> int -> bool

val description : context Storage_description.t

val record_non_consensus_operation_hash : context -> Tezos_protocol_environment_014_PtKathma.Operation_hash.t -> context

val non_consensus_operations : context -> Tezos_protocol_environment_014_PtKathma.Operation_hash.t list

val finalize : ?commit_message:string -> context -> Fitness.raw -> Tezos_protocol_environment_014_PtKathma.Updater.validation_result

Finalize an Alpha_context.t, producing a validation_result.

val current_context : context -> Tezos_protocol_environment_014_PtKathma.Context.t

Should only be used by Main.current_context to return a context usable for RPCs

module Parameters : sig ... end

This module re-exports definitions from Parameters_repr.

module Liquidity_baking : sig ... end

This module re-exports definitions from Liquidity_baking_repr and Liquidity_baking_storage.

module Ticket_balance : sig ... end

This module re-exports definitions from Ticket_storage.

module First_level_of_protocol : sig ... end

module Consensus : sig ... end

This module re-exports definitions from Raw_context.Consensus.

module Token : sig ... end

This module re-exports definitions from Token.

module Fees : sig ... end

This module re-exports definitions from Fees_storage.