Module Tezos_raw_protocol_014_PtKathma.Amendment

Amendments and proposals.

Only delegates with at least one roll take part in the amendment procedure. It works as follows:

Proposal period: delegates can submit protocol amendment proposals using the proposal operation. At the end of a proposal period, the proposal with most supporters is selected and we move to an exploration period. If there are no proposals, or a tie between proposals, a new proposal period starts.

Exploration period: delegates can cast votes to test or not the winning proposal using the ballot operation. At the end of an exploration period if participation reaches the quorum and the proposal has a supermajority in favor, we proceed to a cooldown period. Otherwise we go back to a proposal period. In any case, if there is enough participation the quorum is updated.

Cooldown period: business as usual for the main chain. This period is only a time gap between exploration and promotion periods intended to provide the community with extra time to continue testing the new protocol proposal, and start adapting their infrastructure in advance. At the end of the Cooldown period we move to the Promotion period.

Promotion period: delegates can cast votes to promote or not the proposal using the ballot operation. At the end of a promotion period if participation reaches the quorum and the proposal has a supermajority in favor, we move to an adoption period. Otherwise we go back to a proposal period. In any case, if there is enough participation the quorum is updated.

Adoption period: At the end of an adoption period, the proposal is activated as the new protocol.

The current protocol parameters are documented in src/proto_alpha/lib_parameters/default_parameters.ml

In practice, the real constants used are defined in the migration code. In src/proto_alpha/lib_protocol/init_storage.ml, function prepare_first_block introduces new constants and redefines the existing ones.

val may_start_new_voting_period : Alpha_context.context -> Alpha_context.context Tezos_protocol_environment_014_PtKathma.Error_monad.tzresult Tezos_protocol_environment_014_PtKathma.Lwt.t

If at the end of a voting period, moves to the next one following the state machine of the amendment procedure.

val record_proposals : Alpha_context.context -> Tezos_protocol_environment_014_PtKathma.Chain_id.t -> Alpha_context.public_key_hash -> Tezos_protocol_environment_014_PtKathma.Protocol_hash.t list -> Alpha_context.context Tezos_protocol_environment_014_PtKathma.Error_monad.tzresult Tezos_protocol_environment_014_PtKathma.Lwt.t

Records a list of proposals for a delegate.

raises Unexpected_proposal
if ctxt is not in a proposal period.

raises Unauthorized_proposal
if delegate is not in the listing.

type Tezos_protocol_environment_014_PtKathma.Error_monad.error +=

| Invalid_proposal

| Unexpected_ballot

| Unauthorized_ballot

| Duplicate_ballot

val record_ballot : Alpha_context.context -> Alpha_context.public_key_hash -> Tezos_protocol_environment_014_PtKathma.Protocol_hash.t -> Alpha_context.Vote.ballot -> Alpha_context.context Tezos_protocol_environment_014_PtKathma.Error_monad.tzresult Tezos_protocol_environment_014_PtKathma.Lwt.t

Records a vote for a delegate if the current voting period is Exploration or Promotion.

raises Invalid_proposal
if proposal ≠ current_proposal.

raises Duplicate_ballot
if delegate already voted.

raises Unauthorized_ballot
if delegate is not listed to vote, or if current period differs from Exploration or Promotion.