Module Tezos_raw_protocol_014_PtKathma.Bitset

type t

A bitset is a compact structure to store a set of integers.

type Tezos_protocol_environment_014_PtKathma.Error_monad.error +=

| Invalid_position of int

val encoding : t Tezos_protocol_environment_014_PtKathma.Data_encoding.t

val empty : t

A bitset encoding the empty set.

val mem : t -> int -> bool Tezos_protocol_environment_014_PtKathma.Error_monad.tzresult

mem field i returns true iff i has been added in field.

This functions returns Invalid_input i if i is negative.

val add : t -> int -> t Tezos_protocol_environment_014_PtKathma.Error_monad.tzresult

add field i returns a new bitset which contains i in addition to the previous integers of field.

This functions returns Invalid_input i if i is negative.

val occupied_size_in_bits : t -> int

occupied_size_in_bits bitset returns the current number of bits occupied by the bitset.