Module Tezos_raw_protocol_014_PtKathma.Block_payload_repr

Value on which validators try to reach a consensus.

Consensus at a given level is reached on a sequence of operations. However, to differentiate between two blocks having the same sequence of operations, assuming that could ever happen (for instance, two empty blocks), we also include the hash of the block that precedes the block where these operations should be included.

val hash : predecessor:Tezos_protocol_environment_014_PtKathma.Block_hash.t -> Round_repr.t -> Tezos_protocol_environment_014_PtKathma.Operation_list_hash.t -> Block_payload_hash.t

hash ~predecessor:block_hash round oplh creates a payload hash given a block_hash, the first round at which the payload was proposed and the hash oplh of the non-consensus operations.