Module Tezos_raw_protocol_014_PtKathma.Destination_repr

The type of the destination argument of the Operation_repr.manager_operation.Transaction manager operation.

The introduction of this type allows to interact with emerging layer-2 solutions using the API Tezos users and tooling are already used to: contract calls to entrypoint. These solutions cannot be integrated to Contract_repr.t directly, because values of this type are given a balance, which has an impact on the delegation system.

type t =

| Contract of Contract_repr.t

| Tx_rollup of Tx_rollup_repr.t

| Sc_rollup of Sc_rollup_repr.t

This type is a superset of the set of contracts (Contract_repr.t).

Note: It is of key importance that the encoding of this type remains compatible with Contract_repr.encoding, for the introduction to this type to remain transparent from the existing tooling perspective.

include Tezos_protocol_environment_014_PtKathma.Compare.S with type t := t

val (=) : t -> t -> bool

val (<>) : t -> t -> bool

val (<) : t -> t -> bool

val (<=) : t -> t -> bool

val (>=) : t -> t -> bool

val (>) : t -> t -> bool

val compare : t -> t -> int

val equal : t -> t -> bool

val max : t -> t -> t

val min : t -> t -> t

val to_b58check : t -> string

val of_b58check : string -> t Tezos_protocol_environment_014_PtKathma.Error_monad.tzresult

val encoding : t Tezos_protocol_environment_014_PtKathma.Data_encoding.t

val pp : Tezos_protocol_environment_014_PtKathma.Format.formatter -> t -> unit

val in_memory_size : t -> Cache_memory_helpers.sint

in_memory_size contract returns the number of bytes that are allocated in the RAM for contract.

type Tezos_protocol_environment_014_PtKathma.Error_monad.error +=

| Invalid_destination_b58check of string