Module Tezos_raw_protocol_014_PtKathma.Lazy_storage_kind

Lazy_storage offers a unified interface for specific Michelson datatype that behave somewhat lazily, because they are intended to be quite big. Instead of serializing/deserializing the whole value to/from the storage, only an identifier is used. The identifier acts like a pointer. When using the value in a Michelson script, some part of it may be read from the storage, and a lightweight diff is computed. The diff is effectively applied to the storage at the end of the execution.

This module defines the different kinds of lazy storages and their basic properties. See also Lazy_storage_diff.

Lazy storage types are:

Big_map

module type ID = sig ... end

Lazy storage ids are kept as abstract as possible to avoid mixing them up.

module Big_map : sig ... end

module Sapling_state : sig ... end

type ('id, 'alloc, 'updates) t =

| Big_map : (Big_map.Id.t, Big_map.alloc, Big_map.updates) t

| Sapling_state : (Sapling_state.Id.t, Sapling_state.alloc, Sapling_state.updates) t

Kinds of lazy storage. The GADT ensures operations are properly applied to the correct kind.

'id the abstract type for the identifier of the kind. 'alloc is the type used to construct a new value. 'updates is the type used to update a value.

type ex =

| Ex_Kind : (_, _, _) t -> ex

val all : (int * ex) list

type (_, _) cmp =

| Eq : ('a, 'a) cmp

| Neq

val equal : ('i1, 'a1, 'u1) t -> ('i2, 'a2, 'u2) t -> ('i1 * 'a1 * 'u1, 'i2 * 'a2 * 'u2) cmp

type ('i, 'a, 'u) kind = ('i, 'a, 'u) t

module Temp_ids : sig ... end

Type to manage temporary ids. Used only in the context.

module IdSet : sig ... end