Module Sc_rollup_game_repr.V1

type t = {

turn : player;

inbox_snapshot : Sc_rollup_inbox_repr.t;

level : Raw_level_repr.t;

pvm_name : string;

dissection : (Sc_rollup_repr.State_hash.t option * Sc_rollup_tick_repr.t) list;

}

A game state is characterized by:

turn, the player that must provide the next move.

inbox_snapshot, a snapshot of the inbox state at the moment the game is created. This is only used when checking Input_step and Blocked_step proofs; it makes the proofs easier to create--- otherwise they would have a 'moving target' because the actual inbox may be updated continuously.

level, the inbox level of the commitment the game is refuting. This is only used when checking Blocked_step proofs---the proof will show that the next message available in inbox_snapshot is at level, so shouldn't be included in this commitment.

pvm_name identifies the PVM used in this rollup. It is useful to have here so we can check that the proof provided in a refutation is of the correct kind.

dissection, a list of states with tick counts. The current player will specify, in the next move, a tick count that indicates the last of these states that she agrees with.

Invariants: -----------

dissection must contain at least 3 values

the first state hash value in dissection must not be None

inbox_snapshot never changes once the game is created

val opponent : player -> player

Return the other player

val encoding : t Tezos_protocol_environment_014_PtKathma.Data_encoding.t

val pp_dissection : Tezos_protocol_environment_014_PtKathma.Format.formatter -> (Sc_rollup_repr.State_hash.t option * Sc_rollup_tick_repr.t) list -> unit

val pp : Tezos_protocol_environment_014_PtKathma.Format.formatter -> t -> unit