Module Sc_rollup_management_protocol.Internal_for_tests

Function for constructing and encoding inbox_message and outbox_message values. Since Layer 1 only ever consumes outbox_message values and produces inbox_message values, these functions are used for testing only.

val make_transaction : Alpha_context.context -> ('a, _) Script_typed_ir.ty -> parameters:'a -> destination:Contract_hash.t -> entrypoint:Alpha_context.Entrypoint.t -> (transaction * Alpha_context.context) Tezos_protocol_environment_014_PtKathma.Error_monad.tzresult Tezos_protocol_environment_014_PtKathma.Lwt.t

make_transaction ctxt ty ~parameters ~destination ~entrypoint creates a Layer 1 to Layer 2 transaction.

val make_atomic_batch : transaction list -> outbox_message

make_atomic_batch ts creates an atomic batch with the given transactions ts.

val bytes_of_outbox_message : outbox_message -> string Tezos_protocol_environment_014_PtKathma.Error_monad.tzresult

bytes_of_output_message msg encodes the outbox message msg in binary format.

val inbox_message_of_bytes : string -> Alpha_context.Sc_rollup.Inbox.Message.t Tezos_protocol_environment_014_PtKathma.Error_monad.tzresult

inbox_message_of_bytes bs decodes an inbox message from the given bytes bs.