Module Tezos_raw_protocol_014_PtKathma.Sc_rollup_outbox_message_repr

This module defines a data type t that represents messages from Layer 2 to Layer 1.

They are part of the Rollup Management Protocol that defines the communication protocol for exchanging messages between Layer 1 and Layer 2 for smart-contract rollups.

An outbox-message consists of a sequence of transactions to L1 smart-contract accounts. All transactions contained in a message are intended to be executed as a batch.

type transaction = {

unparsed_parameters : Script_repr.expr;
(*
The payload.
*)

destination : Contract_hash.t;
(*
The recipient contract.
*)

entrypoint : Entrypoint_repr.t;
(*
Entrypoint of the destination.
*)

}

A transaction from L2 to L1.

type t =

| Atomic_transaction_batch of {

transactions : transaction list;

}

A type representing messages from Layer 2 to Layer 1.

val encoding : t Tezos_protocol_environment_014_PtKathma.Data_encoding.t

val pp : Tezos_protocol_environment_014_PtKathma.Format.formatter -> t -> unit

val of_bytes : string -> t Tezos_protocol_environment_014_PtKathma.Error_monad.tzresult

of_bytes ctxt bs decodes an outbox message value from the given bytes bs. The function involves parsing Micheline expressions to typed values.

module Internal_for_tests : sig ... end

Module containing functions exposed so they can be used in test.