Module Tezos_raw_protocol_014_PtKathma.Sc_rollup_storage

val originate : Raw_context.t -> kind:Sc_rollups.Kind.t -> boot_sector:string -> parameters_ty:Script_repr.lazy_expr -> (Sc_rollup_repr.Address.t * Tezos_protocol_environment_014_PtKathma.Z.t * Raw_context.t) Tezos_protocol_environment_014_PtKathma.Error_monad.tzresult Tezos_protocol_environment_014_PtKathma.Lwt.t

originate context ~kind ~boot_sector produces an address a for a smart contract rollup using the origination nonce found in context. This function also initializes the storage with a new entry indexed by a to remember the kind of the rollup at address a and also to remember its boot_sector.

Also returns the number of allocated bytes.

val kind : Raw_context.t -> Sc_rollup_repr.t -> Sc_rollups.Kind.t option Tezos_protocol_environment_014_PtKathma.Error_monad.tzresult Tezos_protocol_environment_014_PtKathma.Lwt.t

kind context address returns Some kind iff address is an existing rollup of some kind. Returns None if address is not the address of an existing rollup.

val list : Raw_context.t -> Sc_rollup_repr.t list Tezos_protocol_environment_014_PtKathma.Error_monad.tzresult Tezos_protocol_environment_014_PtKathma.Lwt.t

val initial_level : Raw_context.t -> Sc_rollup_repr.t -> Raw_level_repr.t Tezos_protocol_environment_014_PtKathma.Error_monad.tzresult Tezos_protocol_environment_014_PtKathma.Lwt.t

initial_level ctxt sc_rollup returns the level at which a sc_rollup was originated.

val get_boot_sector : Raw_context.t -> Sc_rollup_repr.t -> string Tezos_protocol_environment_014_PtKathma.Error_monad.tzresult Tezos_protocol_environment_014_PtKathma.Lwt.t

get_boot_sector ctxt sc_rollup retrieves the boot sector for sc_rollup.

val parameters_type : Raw_context.t -> Sc_rollup_repr.t -> (Script_repr.lazy_expr option * Raw_context.t) Tezos_protocol_environment_014_PtKathma.Error_monad.tzresult Tezos_protocol_environment_014_PtKathma.Lwt.t

parameters_type ctxt rollup returns the registered type of a rollup. Returns None in case there is no registered type for the rollup.

module Outbox : sig ... end

A module for managing state concerning a rollup's outbox.

module Dal_slot : sig ... end