Module Sc_rollups.Kind

A smart contract rollup has a kind, which assigns meaning to rollup operations.

type t =

| Example_arith

| Wasm_2_0_0

The list of available rollup kinds.

This list must only be appended for backward compatibility.

val encoding : t Tezos_protocol_environment_014_PtKathma.Data_encoding.t

val equal : t -> t -> bool

val pp : Tezos_protocol_environment_014_PtKathma.Format.formatter -> t -> unit

val pvm_of : t -> PVM.t

pvm_of kind returns the PVM of the given kind.

val of_pvm : PVM.t -> t

of_pvm pvm returns the kind of the given PVM.

val pvm_of_name : name:string -> PVM.t option

pvm_of_name ~name is Some (module I) if an implemented PVM called name. This function returns None otherwise.

val all : t list

all returns all implemented PVM.

val all_names : string list

all_names returns all implemented PVM names.

val of_name : string -> t option

of_name name returns the kind of the PVM of the specified name.

val name_of : t -> string

name_of kind returns a human-readable representation of kind.