Module Tezos_raw_protocol_014_PtKathma.Script_ir_annot

val error_unexpected_annot : Alpha_context.Script.location -> 'a list -> unit Tezos_protocol_environment_014_PtKathma.Error_monad.tzresult

returns
an error Unexpected_annotation in the monad the list is not empty.

check_xxx_annot functions below are remains from the past (they were called parse_xxx_annot before). They check that annotations are well-formed and, depending on different contexts, that only the annotations that are expected to be found are present. Hopefully we will relax this property soon.

val check_type_annot : Alpha_context.Script.location -> string list -> unit Tezos_protocol_environment_014_PtKathma.Error_monad.tzresult

Check a type annotation only.

val is_field_annot : Alpha_context.Script.location -> string -> bool Tezos_protocol_environment_014_PtKathma.Error_monad.tzresult

Check a field annotation only.

val check_composed_type_annot : Alpha_context.Script.location -> string list -> unit Tezos_protocol_environment_014_PtKathma.Error_monad.tzresult

Check an annotation for composed types, of the form :ty_name %field1 %field2 in any order.

val has_field_annot : Alpha_context.Script.node -> bool Tezos_protocol_environment_014_PtKathma.Error_monad.tzresult

Checks whether a node has a field annotation.

val remove_field_annot : Alpha_context.Script.node -> Alpha_context.Script.node Tezos_protocol_environment_014_PtKathma.Error_monad.tzresult

Removes a field annotation from a node.

val extract_entrypoint_annot : Alpha_context.Script.node -> (Alpha_context.Script.node * Alpha_context.Entrypoint.t option) Tezos_protocol_environment_014_PtKathma.Error_monad.tzresult

Extract and remove a field annotation as an entrypoint from a node

Instruction annotations parsing

val check_var_annot : Alpha_context.Script.location -> string list -> unit Tezos_protocol_environment_014_PtKathma.Error_monad.tzresult

Check a variable annotation.

val is_allowed_char : char -> bool

val check_constr_annot : Alpha_context.Script.location -> string list -> unit Tezos_protocol_environment_014_PtKathma.Error_monad.tzresult

val check_two_var_annot : Alpha_context.Script.location -> string list -> unit Tezos_protocol_environment_014_PtKathma.Error_monad.tzresult

val check_destr_annot : Alpha_context.Script.location -> string list -> unit Tezos_protocol_environment_014_PtKathma.Error_monad.tzresult

val check_unpair_annot : Alpha_context.Script.location -> string list -> unit Tezos_protocol_environment_014_PtKathma.Error_monad.tzresult

val parse_entrypoint_annot_strict : Alpha_context.Script.location -> string list -> Alpha_context.Entrypoint.t Tezos_protocol_environment_014_PtKathma.Error_monad.tzresult

Parses a field annotation and converts it to an entrypoint. An error is returned if the annotation is too long or is "default". An empty annotation is converted to "default".

val parse_entrypoint_annot_lax : Alpha_context.Script.location -> string list -> Alpha_context.Entrypoint.t Tezos_protocol_environment_014_PtKathma.Error_monad.tzresult

Parse a field annotation and convert it to an entrypoint. An error is returned if the field annot is too long. An empty annotation is converted to "default".

val check_var_type_annot : Alpha_context.Script.location -> string list -> unit Tezos_protocol_environment_014_PtKathma.Error_monad.tzresult