Module Tezos_raw_protocol_014_PtKathma.Tx_rollup_commitment_repr

module Hash : sig ... end

A specialized Blake2B implementation for hashing commitments with "toc1" as a base58 prefix

module Merkle_hash : Tezos_protocol_environment_014_PtKathma.S.HASH

module Merkle : Merkle_list.T with type elt = Tx_rollup_message_result_hash_repr.t and type h = Merkle_hash.t

type 'a template = {

level : Tx_rollup_level_repr.t;

messages : 'a;

predecessor : Hash.t option;

inbox_merkle_root : Tx_rollup_inbox_repr.Merkle.root;

}

A commitment describes the interpretation of the messages stored in the inbox of a particular level, on top of a particular layer-2 context.

It includes one Merkle tree root for each of the batches. It has a predecessor, which is the identifier of the commitment for the previous inbox. The predecessor is used to get the Merkle root of the layer-2 context before any inboxes are processed. If predecessor is None, the commitment is for the first inbox with messages in this rollup, and the initial Merkle root is the empty tree.

module Compact : sig ... end

module Full : sig ... end

module Index : Storage_description.INDEX with type t = Hash.t

module Submitted_commitment : sig ... end