Module Tx_rollup_l2_verifier.Verifier_storage

include Tx_rollup_l2_storage_sig.STORAGE with type t = Tezos_protocol_environment_014_PtKathma.Context.tree and type 'a m = ('a, Tezos_protocol_environment_014_PtKathma.Error_monad.error) Tezos_protocol_environment_014_PtKathma.Pervasives.result Tezos_protocol_environment_014_PtKathma.Lwt.t

type t = Tezos_protocol_environment_014_PtKathma.Context.tree

The state of the storage.

The API adopts a functional paradigm, where the set function returns a new state for the storage, and where it should be possible to reuse a previous state.

type 'a m = ('a, Tezos_protocol_environment_014_PtKathma.Error_monad.error) Tezos_protocol_environment_014_PtKathma.Pervasives.result Tezos_protocol_environment_014_PtKathma.Lwt.t

The monad of the storage backend.

module Syntax : sig ... end

The necessary monadic operators the monad of the storage backend is required to provide.

val get : t -> bytes -> bytes option m

get storage key returns the value stored in storage for key, if it exists. Returns None if it does not.

val set : t -> bytes -> bytes -> t m

set storage key computes a new state for the storage wherein the value associated to key is value.

storage is expected to remain usable and consistent even after the execution of set.

val remove : t -> bytes -> t m

remove storage key removes key from the storage.