Module Validate_operation.TMP_for_plugin

Functions for the plugin.

These functions are temporary.

TODO: https://gitlab.com/tezos/tezos/-/issues/3245 Update the plugin to call directly validate_operation then remove these functions.

type 'a should_check_signature =

| Check_signature of 'a Alpha_context.operation

| Skip_signature_check

Indicate whether the signature should be checked in precheck_manager; if so, provide the raw operation.

We could have used an option, but this makes calls to precheck_manager more readable.

val precheck_manager : validate_operation_info -> validate_operation_state -> 'a Alpha_context.Kind.manager Alpha_context.contents_list -> 'a Alpha_context.Kind.manager should_check_signature -> stamp Tezos_protocol_environment_014_PtKathma.Error_monad.tzresult Tezos_protocol_environment_014_PtKathma.Lwt.t

Similar to validate_operation, but do not check the one-operation-per-manager-per-block restriction (1M).

Indeed, 1M is already handled by the plugin. This function is purposefully close to the former Apply.precheck_manager_contents_list, so that few changes are needed in the plugin.

The signature is only checked if the should_check_signature argument is Check_signature _.

The validate_operation_state does not need to be updated because:

1M is not handled here anyway.

In mempool mode, the block gas limit is not tracked.

This function is called by Main.precheck_manager, which is called in lib_plugin/mempool.ml.

val precheck_manager__do_nothing_on_non_manager_op : Alpha_context.t -> Tezos_protocol_environment_014_PtKathma.Chain_id.t -> 'kind Alpha_context.contents_list -> 'kind should_check_signature -> stamp Tezos_protocol_environment_014_PtKathma.Error_monad.tzresult Tezos_protocol_environment_014_PtKathma.Lwt.t

Same as precheck_manager, except that:

This function does not require validate_operation_info and validate_operation_state arguments. Instead, they are constructed internally from the given context and chain_id.

This function accepts any kind of operation as its contents_list argument rather than just manager operations. However, on non-manager operations, this function does not check anything.

This function is called in lib_plugin/RPC.ml, where we do not have access to a Main.validation_state containing validate_operation_info and _state, and where we need a stamp even for non-manager operations.